+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 Noteable/major changes will be listed here
 
+## 10th March 2021
+
+### Added
+
+- verifyAndCorrectGraspPose() method.
+- srv to reset arm using motion planner, this will be used instead of the goToInit() method in the Controller class.
+  
+### Removed
+
+- removed attempt to push objects in opposite direction.
+
+### Fixed
+
+- blocking objects were not saved in the correct variable.
+
+## 7th March 2021
+
+### Added
+
+- if init state is not valid, an offset is applied to the y-value as this mostly happens when the target object is near a wall
+
+### Modified
+
+- isObjectBlocked() now takes in a vector of collision geometries and sorts them instead of doing this inside planInClutter()
+
 ## 5th March 2021
 
 ### Fixed